`Golden`: add examples for globals that indirectly reference untagged types
This shows that there is a bug: we try to pretty-print such types, but we can't
because there is no C name to refer to for untagged types.
